+++ ui/src/components/CreateMewField.vue
@@ -20,14 +20,25 @@
           self="bottom middle"
           :delay="TOOLTIP_DELAY"
         >
-          You can mention people with @ and use #hashtags and $cashtags in a
-          mew.
+          You can mention people with @ and use #hashtags and $cashtags as well
+          as ^links in a mew.
         </q-tooltip>
       </q-icon>
 
       <div class="q-mb-xs text-right text-caption text-grey">
         Ctrl/Cmd + Enter to publish
       </div>
+
+      <q-card class="link-text q-px-md q-py-sm" style="min-width: 13rem">
+        <q-input
+          v-model="linkText"
+          type="text"
+          placeholder="Link description (optional)"
+          dense
+          borderless
+          @keyup="onLinkTextKeyDown"
+        />
+      </q-card>
 
       <q-card class="autocompleter">
         <template v-if="currentAgentSearch.length < 3">
@@ -61,9 +72,8 @@
                 <agent-avatar :agentPubKey="agent.key" size="30" />
               </q-item-section>
               <q-item-section>
-                {{ agent.value.fields[PROFILE_FIELDS.DISPLAY_NAME] }} @{{
-                  agent.value.nickname
-                }}
+                {{ agent.value.fields[PROFILE_FIELDS.DISPLAY_NAME] }}
+                {{ TAG_SYMBOLS.MENTION }}{{ agent.value.nickname }}
               </q-item-section>
             </q-item>
 
@@ -91,6 +101,7 @@
 import { useProfilesStore } from "@/services/profiles-store";
 import { useClutterStore } from "@/stores";
 import { showError } from "@/utils/notification";
+import { TAG_SYMBOLS } from "@/utils/tags";
 import { Profile } from "@holochain-open-dev/profiles";
 import { debounce } from "quasar";
 import { onMounted, PropType, ref } from "vue";
@@ -104,6 +115,7 @@ import {
 } from "../types/types";
 
 const ANCHOR_DATA_ID_AGENT_PUB_KEY = "agentPubKey";
+const ANCHOR_DATA_ID_URL = "url";
 
 const emit = defineEmits<{ (e: "publish-mew"): void }>();
 
@@ -125,12 +137,14 @@ onMounted(() => setTimeout(focusInputField, 0));
 const isMewEmpty = ref(true);
 const saving = ref(false);
 
+const linkText = ref("");
+
 const currentAgentSearch = ref("");
 let currentAnchorOffset: number;
 let currentFocusOffset: number;
 let currentNode: Node;
 
-const AUTOCOMPLETER_MARGIN_TOP = 20;
+const POPUP_MARGIN_TOP = 20;
 const agentAutocompletions = ref<AgentAutocompletion[]>([]);
 const autocompleterLoading = ref(false);
 
@@ -139,10 +153,10 @@ const focusInputField = () =>
     .getSelection()
     ?.setPosition(mewContainer.value?.querySelector(".mew-content") || null, 0);
 
-const hideAutocompleter = () => {
-  const autocompleter = mewContainer.value?.querySelector(".autocompleter");
-  if (autocompleter && autocompleter instanceof HTMLElement) {
-    autocompleter.style.display = "none";
+const hideElement = (selector: string) => {
+  const element = mewContainer.value?.querySelector(selector);
+  if (element && element instanceof HTMLElement) {
+    element.style.display = "none";
   }
 };
 
@@ -188,6 +202,7 @@ const onKeyUp = (keyUpEvent: KeyboardEvent) => {
   const selection = document.getSelection();
   if (keyUpEvent.key === " ") {
     hideAutocompleter();
+    hideLinkTextInput();
   } else if (
     (keyUpEvent.key === "Backspace" || keyUpEvent.key === "Delete") &&
     selection?.anchorNode?.parentElement?.tagName === "A"
@@ -199,7 +214,10 @@ const onKeyUp = (keyUpEvent: KeyboardEvent) => {
   }
 };
 
-const onMouseUp = () => hideAutocompleter();
+const onMouseUp = () => {
+  hideAutocompleter();
+  hideLinkTextInput();
+};
 
 const onPaste = (event: ClipboardEvent) => {
   event.preventDefault();
@@ -208,6 +226,8 @@ const onPaste = (event: ClipboardEvent) => {
     const pastedNode = document.createTextNode(data);
     document.getSelection()?.getRangeAt(0).insertNode(pastedNode);
     document.getSelection()?.setPosition(pastedNode, pastedNode.length);
+    onCaretPositionChange();
+    onInput();
   }
 };
 
@@ -222,6 +242,30 @@ const loadAutocompleterUsers = async (nickname: string) => {
     showError(error);
   } finally {
     autocompleterLoading.value = false;
+  }
+};
+
+const onLinkTextKeyDown = (keyDownEvent: KeyboardEvent) => {
+  if (keyDownEvent.key === "Enter") {
+    const range = new Range();
+    range.setStart(currentNode, currentAnchorOffset);
+    range.setEnd(currentNode, currentFocusOffset);
+    const url = range.extractContents().textContent;
+
+    const anchor = document.createElement("a");
+    anchor.href = "#";
+    anchor.textContent = linkText.value
+      ? TAG_SYMBOLS.URL + linkText.value
+      : url;
+    anchor.dataset[ANCHOR_DATA_ID_URL] = url ?? undefined;
+    range.insertNode(anchor);
+    // insert space after link
+    const spaceNode = document.createTextNode(String.fromCharCode(160));
+    anchor.after(spaceNode);
+
+    hideLinkTextInput();
+    document.getSelection()?.setPosition(spaceNode, 1);
+    linkText.value = "";
   }
 };
 
@@ -258,7 +302,7 @@ const onAutocompleteAgentSelect = (agent: AgentAutocompletion) => {
   range.setEnd(currentNode, currentFocusOffset);
   const anchor = document.createElement("a");
   anchor.href = "#";
-  anchor.textContent = "@" + agent.value.nickname;
+  anchor.textContent = TAG_SYMBOLS.MENTION + agent.value.nickname;
   anchor.dataset[ANCHOR_DATA_ID_AGENT_PUB_KEY] = agent.key.toString();
   range.deleteContents();
   range.insertNode(anchor);
@@ -276,28 +320,32 @@ const publishMew = async () => {
   }
 
   // build link array
-  const mentions: LinkTarget[] = [];
+  const links: LinkTarget[] = [];
   for (let i = 0; i < mewInput.children.length; i++) {
     const element = mewInput.children.item(i);
     if (
       element &&
       element.tagName === "A" &&
-      element instanceof HTMLAnchorElement &&
-      element.dataset[ANCHOR_DATA_ID_AGENT_PUB_KEY]
+      element instanceof HTMLAnchorElement
     ) {
-      const agentPubKeyString = element.dataset[ANCHOR_DATA_ID_AGENT_PUB_KEY];
-      // dunno why this is an error; an array is array-like
-      // eslint-disable-next-line
-      // @ts-ignore
-      const agentPubKey = Uint8Array.from(agentPubKeyString.split(","));
-      mentions.push({ [LinkTargetName.Mention]: agentPubKey });
+      if (element.dataset[ANCHOR_DATA_ID_AGENT_PUB_KEY]) {
+        const agentPubKeyString = element.dataset[ANCHOR_DATA_ID_AGENT_PUB_KEY];
+        // dunno why this is an error; an array is array-like
+        // eslint-disable-next-line
+        // @ts-ignore
+        const agentPubKey = Uint8Array.from(agentPubKeyString.split(","));
+        links.push({ [LinkTargetName.Mention]: agentPubKey });
+      } else if (element.dataset[ANCHOR_DATA_ID_URL]) {
+        const url = element.dataset[ANCHOR_DATA_ID_URL];
+        links.push({ [LinkTargetName.URL]: url });
+      }
     }
   }
 
   const createMewInput: CreateMewInput = {
     mewType: props.mewType,
     text: mewInput.textContent ? mewInput.textContent.trim() : null,
-    links: mentions.length ? mentions : undefined,
+    links: links.length ? links : undefined,
   };
   try {
     saving.value = true;
@@ -328,7 +376,7 @@ const onCaretPositionChange = () => {
   const endOfWordIndex = selection.anchorOffset + endOfAheadIndex;
 
   // find start of word that the caret is positioned at
-  const behind = content.substring(0, selection.anchorOffset);
+  const behind = content.substring(0, selection.anchorOffset - 1);
   let lastSpaceIndex = -1;
   // find last index of space, which can be " " (32) or "&nbsp;" (160)
   for (let i = behind.length - 1; i >= 0; i--) {
@@ -341,32 +389,10 @@ const onCaretPositionChange = () => {
   const startOfWordIndex = lastSpaceIndex === -1 ? 0 : lastSpaceIndex + 1;
   const currentWord = content.substring(startOfWordIndex, endOfWordIndex);
 
-  if (
-    currentWord.length &&
-    currentWord.startsWith("@") &&
-    /^@\w+/.test(currentWord) &&
-    selection.anchorNode
-  ) {
+  if (currentWord.length && selection.anchorNode) {
     // current word starts with @ and is followed by at least another word character
-    const range = new Range();
-    range.setStart(selection.anchorNode, startOfWordIndex);
-    currentNode = selection.anchorNode;
-    const selectionRect = range.getBoundingClientRect();
-    const autocompleter = mewContainer.value?.querySelector(".autocompleter");
-    if (autocompleter instanceof HTMLElement && mewContainer.value) {
-      autocompleter.style.top =
-        Math.max(
-          0,
-          selectionRect.top - mewContainer.value.getBoundingClientRect().top
-        ) +
-        AUTOCOMPLETER_MARGIN_TOP +
-        "px";
-      autocompleter.style.left =
-        Math.max(
-          0,
-          selectionRect.left - mewContainer.value.getBoundingClientRect().left
-        ) + "px";
-      autocompleter.style.display = "block";
+    if (new RegExp(`^${TAG_SYMBOLS.MENTION}\\w+`).test(currentWord)) {
+      showElement(selection.anchorNode, startOfWordIndex, ".autocompleter");
 
       const nicknameChars = currentWord.substring(1);
       currentAgentSearch.value = nicknameChars;
@@ -375,12 +401,50 @@ const onCaretPositionChange = () => {
         currentAnchorOffset = startOfWordIndex;
         currentFocusOffset = endOfWordIndex;
       }
+      // current word is a URL
+    } else if (URL_REGEX.test(currentWord)) {
+      showElement(selection.anchorNode, startOfWordIndex, ".link-text");
+      currentAnchorOffset = startOfWordIndex;
+      currentFocusOffset = endOfWordIndex;
+    } else {
+      hideAutocompleter();
+      hideLinkTextInput();
     }
-  } else {
-    hideAutocompleter();
   }
 };
 const debouncedOnCaretPositionChange = debounce(onCaretPositionChange, 300);
+
+const showElement = (
+  anchorNode: Node,
+  startOfWordIndex: number,
+  selector: string
+) => {
+  const range = new Range();
+  range.setStart(anchorNode, startOfWordIndex);
+  currentNode = anchorNode;
+  const selectionRect = range.getBoundingClientRect();
+  const element = mewContainer.value?.querySelector(selector);
+  if (element instanceof HTMLElement && mewContainer.value) {
+    element.style.top =
+      Math.max(
+        0,
+        selectionRect.top - mewContainer.value.getBoundingClientRect().top
+      ) +
+      POPUP_MARGIN_TOP +
+      "px";
+    element.style.left =
+      Math.max(
+        0,
+        selectionRect.left - mewContainer.value.getBoundingClientRect().left
+      ) + "px";
+    element.style.display = "block";
+  }
+};
+const hideAutocompleter = hideElement.bind(null, ".autocompleter");
+const hideLinkTextInput = hideElement.bind(null, ".link-text");
+const URL_REGEX = new RegExp(
+  `^(http[s]?:\\/\\/(www\\.)?|www\\.){1}([0-9A-Za-z-\\.@:%_\\+~#=]+)+((\\.[a-zA-Z]{2,3})+)(/(.)*)?(\\?(.)*)?`
+);
 </script>
 
 <style lang="sass">
@@ -399,7 +463,7 @@ const debouncedOnCaretPositionChange = debounce(onCaretPositionChange, 300);
   top: 5px
   right: 5px
 
-.autocompleter
+.autocompleter, .link-text
   position: absolute
   display: none
   z-index: 1