+++ ui/src/components/MewContent.vue
@@ -2,16 +2,27 @@
   <div class="column">
     <div class="text-body1" style="overflow-wrap: anywhere">
       <span v-for="(contentPart, index) of contentParts" :key="index">
-        <router-link
-          v-if="contentPart[1]"
-          :to="contentPart[1]"
-          class="text-secondary text-bold"
-          @click.stop
-        >
-          {{ contentPart[0] }}
-        </router-link>
+        <template v-if="Array.isArray(contentPart)">
+          <a
+            v-if="typeof contentPart[1] === 'string'"
+            :href="contentPart[1]"
+            target="_blank"
+            @click.stop
+          >
+            {{ contentPart[0] }}
+          </a>
 
-        <template v-else>{{ contentPart[0] }}</template>
+          <router-link
+            v-else
+            :to="contentPart[1]"
+            class="text-secondary text-bold"
+            @click.stop
+          >
+            {{ contentPart[0] }}
+          </router-link>
+        </template>
+
+        <template v-else>{{ contentPart }}</template>
       </span>
     </div>
   </div>
@@ -25,7 +36,7 @@ import { serializeHash } from "@holochain-open-dev/utils";
 import { computed, PropType } from "vue";
 import { RouteLocationRaw } from "vue-router";
 
-type ContentPart = [string] | [string, RouteLocationRaw];
+type ContentPart = string | [string, RouteLocationRaw] | [string, string];
 
 const props = defineProps({
   feedMew: { type: Object as PropType<FeedMew>, required: true },
@@ -34,7 +45,7 @@ const props = defineProps({
 const startsWithTag = (contentPart: string) =>
   Object.values(TAG_SYMBOLS).some((symbol) => contentPart.startsWith(symbol));
 
-const mentions = computed(() =>
+const links = computed(() =>
   props.feedMew.mew.content?.links?.slice().reverse()
 );
 const content = computed(() => props.feedMew.mew.content?.text || "");
@@ -45,20 +56,34 @@ const parts = computed(() =>
 const contentParts = computed<ContentPart[]>(() =>
   parts.value.map((part) => {
     if (startsWithTag(part)) {
-      let agentPubKey = "";
-      if (part[0] === TAG_SYMBOLS.MENTION) {
-        const mention = mentions.value?.pop()?.[LinkTargetName.Mention];
-        // formality, mention must exist
-        agentPubKey = mention ? serializeHash(mention) : "";
+      let agentPubKey: string | undefined = undefined;
+      if (part[0] === TAG_SYMBOLS.MENTION || part[0] === TAG_SYMBOLS.URL) {
+        const link = links.value?.pop();
+        if (!link) {
+          return part;
+        }
+        if (LinkTargetName.Mention in link) {
+          const mention = link[LinkTargetName.Mention];
+          agentPubKey = serializeHash(mention);
+        } else if (LinkTargetName.URL in link) {
+          const url = link[LinkTargetName.URL];
+          return [part, url];
+        }
       }
       const to: RouteLocationRaw = {
         name: ROUTES[PATH[part[0]]],
         params: { tag: part.substring(1) },
-        query: { agentPubKey: agentPubKey || undefined },
+        query: { agentPubKey },
       };
       return [part, to];
     }
-    return [part];
+    return part;
   })
 );
 </script>
+
+<style lang="sass">
+a
+    color: $secondary
+    font-weight: 600
+</style>