+++ ui/src/types/types.ts
@@ -10,11 +10,16 @@ export const TOOLTIP_DELAY = 400;
 
 export enum LinkTargetName {
   Mention = "Mention",
+  URL = "URL",
 }
 
-export type LinkTarget = {
-  [LinkTargetName.Mention]: AgentPubKey;
-};
+export type LinkTarget =
+  | {
+      [LinkTargetName.Mention]: AgentPubKey;
+    }
+  | {
+      [LinkTargetName.URL]: string;
+    };
 
 export type CreateMewInput = {
   mewType: MewType;
@@ -86,4 +91,10 @@ export interface NotificationOptions {
     | "center";
   actions?: Array<() => void>;
   onDismiss?: () => void;
+}
+
+export enum SearchResult {
+  Agent,
+  Hashtag,
+  Cashtag,
 }